# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Backend Assignment 2 API",
    "version": "1.0.0",
    "description": "API documentation for the backend assignment"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
      "Branch": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the branch"
          },
          "name": {
            "type": "string",
            "description": "Name of the branch"
          },
          "address": {
            "type": "string",
            "description": "Address of the branch"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the branch"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "phone"
        ]
      },
      "BranchCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the branch"
          },
          "address": {
            "type": "string",
            "description": "Address of the branch"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the branch"
          }
        },
        "required": [
          "name",
          "address",
          "phone"
        ]
      },
      "BranchUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the branch"
          },
          "address": {
            "type": "string",
            "description": "Address of the branch"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the branch"
          }
        }
      },
      "Employee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the employee"
          },
          "name": {
            "type": "string",
            "description": "Name of the employee"
          },
          "position": {
            "type": "string",
            "description": "Position of the employee"
          },
          "department": {
            "type": "string",
            "description": "Department of the employee"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email of the employee"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the employee"
          },
          "branchId": {
            "type": "string",
            "description": "ID of the branch the employee belongs to"
          }
        },
        "required": [
          "id",
          "name",
          "position",
          "department",
          "email",
          "phone",
          "branchId"
        ]
      },
      "EmployeeCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the employee"
          },
          "position": {
            "type": "string",
            "description": "Position of the employee"
          },
          "department": {
            "type": "string",
            "description": "Department of the employee"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email of the employee"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the employee"
          },
          "branchId": {
            "type": "string",
            "description": "ID of the branch the employee belongs to"
          }
        },
        "required": [
          "name",
          "position",
          "department",
          "email",
          "phone",
          "branchId"
        ]
      },
      "EmployeeUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the employee"
          },
          "position": {
            "type": "string",
            "description": "Position of the employee"
          },
          "department": {
            "type": "string",
            "description": "Department of the employee"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email of the employee"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the employee"
          },
          "branchId": {
            "type": "string",
            "description": "ID of the branch the employee belongs to"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status of the response"
          },
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "message": {
            "type": "string",
            "description": "Additional message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          }
        }
      }
    }
  },
  "paths": {
    "/api/v1/branches": {
      "get": {
        "summary": "Retrieve all branches",
        "tags": [
          "Branches"
        ],
        "responses": {
          "200": {
            "description": "A list of branches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Branch"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "post": {
        "summary": "Create a new branch",
        "tags": [
          "Branches"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Branch created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
    "/api/v1/branches/{id}": {
      "get": {
        "summary": "Get a branch by ID",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Branch unique ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Branch found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "put": {
        "summary": "Update an existing branch",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Branch updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/Error"
          },
          "404": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "delete": {
        "summary": "Delete a branch",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Branch deleted successfully (no content)"
          },
          "404": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
    "/api/v1/employees": {
      "get": {
        "summary": "Retrieve a list of employees",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Maximum number of employees to return"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of employees",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Employee"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "post": {
        "summary": "Create a new employee",
        "tags": [
          "Employees"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
    "/api/v1/employees/{id}": {
      "get": {
        "summary": "Get an employee by ID",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "put": {
        "summary": "Update an employee",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/Error"
          },
          "404": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "delete": {
        "summary": "Delete an employee",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          },
          "404": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
    "/api/v1/branches/{branchId}/employees": {
      "get": {
        "summary": "Get employees for a specific branch",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "branchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employees in branch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/departments/{department}/employees": {
      "get": {
        "summary": "Get employees in a specific department",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "department",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employees in department",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}